# [10.x] Improve `Arr::dot` performance