--- a/src/Illuminate/Collections/Arr.php
+++ b/src/Illuminate/Collections/Arr.php
@@ -109,17 +109,17 @@ class Arr
      */
     public static function dot($array, $prepend = '')
     {
-        $results = [];
+        $results = [[]];
 
         foreach ($array as $key => $value) {
             if (is_array($value) && ! empty($value)) {
-                $results += static::dot($value, $prepend.$key.'.');
+                $results[] = static::dot($value, $prepend.$key.'.');
             } else {
-                $results[$prepend.$key] = $value;
+                $results[0][$prepend.$key] = $value;
             }
         }
 
-        return $results;
+        return array_merge(...$results);
     }
 
     /**