--- a/tests/Support/SupportArrTest.php
+++ b/tests/Support/SupportArrTest.php
@@ -119,14 +119,14 @@ class SupportArrTest extends TestCase
             'user.name' => 'Taylor',
             'user.age' => 25,
             'user.languages.0' =>'PHP',
-            'user.languages.1' => 'C#'
+            'user.languages.1' => 'C#',
         ], $array);
 
         $array = Arr::dot(['foo', 'foo' => ['bar' => 'baz', 'baz' => ['a' => 'b']]]);
         $this->assertEquals([
             'foo',
             'foo.bar' => 'baz',
-            'foo.baz.a' => 'b'
+            'foo.baz.a' => 'b',
         ], $array);
     }