Re submission of #49380 by @comes 

Original description : 

> This PR improves the speed of `Arr::dot()` by using the array union operator instead of the slower `array_merge`
> 
> The reason why we can use the union operator here is, that an array always has unique keys.
> 
> > the + operator will only add the elements of the right side operand, if their key doesn't exist in the leftside operand, while array_merge will override existing keys.
> 
> see https://stitcher.io/blog/array-merge-vs+
> 
> But this is not a problem here, because, as I already noticed, the keys in an array are unique.
> 
> I've attached some benchmarks running the current Arr::dot() implementation agains my new implementation. As you can see, the new implementation solves a significant runtime issue.
> 
> ```
> Items: 100
> old Arr::dot: 0.0003 seconds / 20.00 memory
> new Arr::dot: 0.0001 seconds / 20.00 memory
> 
> 
> Items: 1000
> old Arr::dot: 0.0218 seconds / 22.00 memory
> new Arr::dot: 0.0010 seconds / 24.00 memory
> 
> 
> Items: 10000
> old Arr::dot: 3.5038 seconds / 46.00 memory
> new Arr::dot: 0.0129 seconds / 48.50 memory
> 
> 
> Items: 15000
> old Arr::dot: 10.4789 seconds / 64.00 memory
> new Arr::dot: 0.0155 seconds / 68.50 memory
> ```
> 
> for futher information about the union operator see: https://www.php.net/manual/en/language.operators.array.php

~I don't add my benchmark results as they are similar to the original ones~

After some tweak and suggestions (thanks @donnysim and @comes), heres an updated version of the benchmark.  

```
Items: 100
original Arr::dot: 0.0003 seconds / 40.50 memory
new Arr::dot: 0.0002 seconds / 40.50 memory
Items: 1000
original Arr::dot: 0.0213 seconds / 42.50 memory
new Arr::dot: 0.0016 seconds / 42.50 memory
Items: 10000
original Arr::dot: 5.0146 seconds / 68.50 memory
new Arr::dot: 0.0201 seconds / 72.50 memory
Items: 15000
original Arr::dot: 12.8689 seconds / 100.50 memory
new Arr::dot: 0.0272 seconds / 97.50 memory
```
Note : do not compare with original benchmark, it was not run on the same machine

I also updated the tests with `assertSame` instead of `assertEquals` to ensure that the order of the arrays are respected.