@jbrooksuk 2023-12-15
I think it'd be useful to prove that this PR doesn't change existing behaviour, with an additional test.