@bastien-phi 2023-12-15
@jbrooksuk I guessed that existing tests were enough but I added some based on Arr::undot tests