@comes 2023-12-15
@bastien-phi, I appreciate your support here.

I've attached my script to reproduce the results with an unmodified version of laravel.

```php
class ArrDotBench extends Command
{
    protected $signature = 'dot:bench';

    public function handle(): void
    {
        foreach ([100, 1_000, 10_000, 15_000] as $itemCount) {
            $data = $this->data($itemCount);

            $this->framework($data);
            $this->newdot($data);
        }

    }

    protected function data(int $itemCount): array
    {
        $subset = [];
        for ($i=0; $i < $itemCount; $i++) {
            $subset[] = [
                'id' => $i,
                'name' => 'Name ' . $i,
                'email' => 'email' . $i . '@example.com',
                'date' => now()->addDays($i)->format('Y-m-d'),
                'another_sub_array' => [
                    'street' => 'Street ' . $i,
                    'city' => 'City ' . $i,
                    'country' => 'Country ' . $i,
                ],
            ];
        }

        $data = ['items' => $subset];

        $this->info('Items: ' . count($data['items']));

        return $data;
    }

    protected function framework(array $data): array
    {
        $start = $this->start();

        $dottedArr = Arr::dot($data);

        $this->finish('framework dot', $start);

        return $dottedArr;
    }

    protected function newdot(array $data): array
    {
        $start = $this->start();

        $dottedArr = $this->fastdot($data);

        $this->finish('fastdot', $start);

        return $dottedArr;
    }

    protected function start(): float
    {
        memory_reset_peak_usage();

        return microtime(true);
    }

    protected function finish(string $title, float $start): void
    {
        $this->info(sprintf($title . ': %.4f seconds / %.2f memory', microtime(true) - $start, memory_get_peak_usage(true) / 1024 / 1024)).PHP_EOL;
    }

    public static function fastdot($array, $prepend = '')
    {
        $results = [];

        foreach ($array as $key => $value) {
            if (is_array($value) && !empty($value)) {
                $results += static::fastdot($value, $prepend . $key . ".");
            } else {
                $results[$prepend . $key] = $value;
            }
        };

        return $results;
    }
}
```