@donnysim 2023-12-15
You can also do the usual array_merge spread, which is a fraction slower than the union operator, but at a way lower memory cost:

```php
    public function spread($array, $prepend = '')
    {
        $results = [[]];

        foreach ($array as $key => $value) {

            if (is_array($value) && !empty($value)) {
                $results[] = $this->spread($value, $prepend . $key . '.');
            } else {
                $results[0][$prepend . $key] = $value;
            }
        }

        return array_merge(...$results);
    }
```

out based on comes bench:
```
Items: 100
framework dot: 0.0007 seconds / 32.00 memory
fastdot: 0.0003 seconds / 32.00 memory
mergespread: 0.0003 seconds / 32.00 memory
Items: 1000
framework dot: 0.0280 seconds / 34.00 memory
fastdot: 0.0026 seconds / 34.00 memory
mergespread: 0.0031 seconds / 34.00 memory
Items: 10000
framework dot: 4.5470 seconds / 60.00 memory
fastdot: 0.0317 seconds / 66.00 memory
mergespread: 0.0378 seconds / 58.00 memory
Items: 15000
framework dot: 12.1578 seconds / 78.00 memory
fastdot: 0.0468 seconds / 80.00 memory
mergespread: 0.0617 seconds / 72.00 memory
```