@comes 2023-12-15
@donnysim awesome, I prefer your solution. based on a benchmark with up to 1M items it's faster and uses less memory.

```
Items: 100
fastdot: 0.0001 seconds / 20.00 memory
spread: 0.0001 seconds / 20.00 memory
Items: 1000
fastdot: 0.0009 seconds / 24.00 memory
spread: 0.0009 seconds / 22.00 memory
Items: 10000
fastdot: 0.0129 seconds / 48.50 memory
spread: 0.0127 seconds / 45.00 memory
Items: 15000
fastdot: 0.0204 seconds / 66.50 memory
spread: 0.0168 seconds / 59.00 memory
Items: 50000
fastdot: 0.0952 seconds / 162.00 memory
spread: 0.0705 seconds / 142.00 memory
Items: 100000
fastdot: 0.1972 seconds / 326.02 memory
spread: 0.1685 seconds / 270.03 memory
Items: 500000
fastdot: 1.7555 seconds / 1286.02 memory
spread: 1.3012 seconds / 1169.78 memory
Items: 1000000
fastdot: 3.7078 seconds / 2878.02 memory
spread: 2.7969 seconds / 2429.53 memory
```