@bastien-phi 2023-12-15
Agree with that, this version is also faster on my benchmarks. I update this PR