@comes 2023-12-15
> Agree with that, this version is also faster on my benchmarks. I update this PR

Thank you for your support and improvement. Like your solution more than mine.