@donnysim 2023-12-15
Oh, i didn't think of ordering, nice catch 👍 it's possible to optimize it a bit more in this case to group attribute batches instead of array per attribute:

```php
    public function spread($array, $prepend = '')
    {
        $results = [];
        $attrs = [];

        foreach ($array as $key => $value) {
            if (is_array($value) && !empty($value)) {
                if ($attrs) {
                    $results[] = $attrs;
                    $attrs = [];
                }

                $results[] = $this->mergespread($value, $prepend . $key . '.');
            } else {
                $attrs[$prepend . $key] = $value;
            }
        }

        if ($attrs) {
            $results[] = $attrs;
        }

        return array_merge(...$results);
    }
```

haven't tested it fully how much it makes a diff, but it should be little bit better performance and memory wise.