@comes 2023-12-15
> ```
> public function spread($array, $prepend = '')
>     {
>         $results = [[]];
> 
>         foreach ($array as $key => $value) {
> 
>             if (is_array($value) && !empty($value)) {
>                 $results[] = $this->spread($value, $prepend . $key . '.');
>             } else {
>                 $results[0][$prepend . $key] = $value;
>             }
>         }
> 
>         return array_merge(...$results);
>     }
> ```

this 'improved' version, did not perform very well:

```
Items: 100
fastdot: 0.0001 seconds / 20.00 memory
dospread: 0.0001 seconds / 20.00 memory
dospreadImproved: 0.0001 seconds / 20.00 memory
Items: 1000
fastdot: 0.0012 seconds / 24.00 memory
dospread: 0.0013 seconds / 22.00 memory
dospreadImproved: 0.0012 seconds / 24.00 memory
Items: 10000
fastdot: 0.0182 seconds / 48.50 memory
dospread: 0.0149 seconds / 45.00 memory
dospreadImproved: 0.0134 seconds / 50.00 memory
Items: 15000
fastdot: 0.0229 seconds / 66.50 memory
dospread: 0.0183 seconds / 59.00 memory
dospreadImproved: 0.0231 seconds / 64.00 memory
Items: 25000
fastdot: 0.0435 seconds / 99.00 memory
dospread: 0.0375 seconds / 84.00 memory
dospreadImproved: 0.0330 seconds / 94.00 memory
Items: 50000
fastdot: 0.0909 seconds / 172.00 memory
dospread: 0.0792 seconds / 146.00 memory
dospreadImproved: 0.1138 seconds / 162.00 memory
Items: 100000
fastdot: 0.2218 seconds / 326.02 memory
dospread: 0.1973 seconds / 270.03 memory
dospreadImproved: 0.2313 seconds / 306.02 memory
```