@bastien-phi 2023-12-15
AFK, I convert this PR to draft and will try your version when I can